remove JulesConfigMeta; develop JulesInput; tests passing
