+++ src/jules_pytk/config/__init__.py
@@ -1,0 +1,3 @@
+from jules_pytk.config.namelists import *
+from jules_pytk.config.inputs import *
+from jules_pytk.config.config import *
